weather_etl/requirements.txt
requests==2.31.0
pandas==2.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.21
schedule==1.2.0
python-dateutil==2.8.2
